requirements.txt
aniso8601==9.0.1
# asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1652752042988/work
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
datetime-truncate==1.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2
django-admin-charts==1.2.4
django-allauth==0.51.0
django-bower==5.2.0
django-elasticsearch-dsl==7.2.2
django-extensions==3.2.1
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
promise==2.3
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
# sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1663990453245/work
text-unidecode==1.3
# typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
Unidecode==1.3.6
urllib3==1.26.13
